fix(security,ux): P2-1, P2-2, P2-6, P2-7, P2-8
P2-1: Runtime validation of catalog items (name, description, category)
P2-2: JSON body size limit (1MB) for order/review endpoints
P2-6: Deduplicate reviews — single renderer in reviews.js
P2-7: Fix figure_hero.py --axis type (float → int)
P2-8: Sync file limits (15MB/file, 40MB total) between client and worker

+++ tools/figure_hero.py
@@ -449,7 +449,7 @@ def main():
     ap = argparse.ArgumentParser()
     ap.add_argument("--in", dest="src", required=True)
     ap.add_argument("--out", dest="dst")
-    ap.add_argument("--axis", type=float, default=None)
+    ap.add_argument("--axis", type=int, default=None)
     ap.add_argument("--no-head", action="store_true")
     ap.add_argument("--no-paint", action="store_true")
     ap.add_argument("--quality", type=int, default=88)
